+++ CHANGELOG.md
@@ -8,6 +8,11 @@ All notable changes to this project will be documented in this file.
 - Added the possibility of specifying a custom hex colour in the highlight setting and the transformation associated with this colour. Added a "{{CustomHex}}" placeholder in the template gathering all the highlight using that colour
 - The backlink now directs to the specific annotation rather than the page in the pdf file (thanks to @CaoKang-David)
 - It is possible to name the notes with the author (thanks to @ThomasFKJorna)
+- Added {{firstNameInitials}} to the placeholders available to format the name of authors/editors. This is in addition to {{firstName}} and {{lastName}}
+
+### Bugs
+- Fixed bugs that prevented notes added in Zotero using markdown from being formatted properly
+
 
 ## [0.9.10]
 ### Improvement